--- a/src/cache_element.jl
+++ b/src/cache_element.jl
@@ -1,8 +1,0 @@
-mutable struct CacheElement{T}
-    data::T
-    dirty::Bool
-    CacheElement(data::T) where {T} = new{T}(data, true)
-end
-
-@inline setdirty!(element::CacheElement) = (element.dirty = true; nothing)
-@inline isdirty(element::CacheElement) = element.dirty