--- a/src/rigid_body.jl
+++ b/src/rigid_body.jl
@@ -38,9 +38,10 @@ RigidBody(inertia::SpatialInertia) = RigidBody(string(inertia.frame), inertia)
 Base.string(b::RigidBody) = b.name
 Base.show(io::IO, b::RigidBody) = print(io, "RigidBody: \"$(string(b))\"")
 Base.showcompact(io::IO, b::RigidBody) = print(io, "$(string(b))")
-RigidBodyDynamics.Graphs.vertex_id_type(::Type{<:RigidBody}) = BodyID
-RigidBodyDynamics.Graphs.vertex_id(b::RigidBody) = b.id
-RigidBodyDynamics.Graphs.set_vertex_id!(b::RigidBody, id::BodyID) = (b.id = id)
+@inline id(b::RigidBody) = b.id
+@inline RigidBodyDynamics.Graphs.vertex_id_type(::Type{<:RigidBody}) = BodyID
+@inline RigidBodyDynamics.Graphs.vertex_id(b::RigidBody) = id(b)
+@inline RigidBodyDynamics.Graphs.set_vertex_id!(b::RigidBody, id::BodyID) = (b.id = id)
 
 """
 $(SIGNATURES)