--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -11,7 +11,7 @@ using ForwardDiff
 import Base.Iterators: filter
 
 # useful utility function for computing time derivatives.
-create_autodiff(x, dx) = [ForwardDiff.Dual(x[i], dx[i]) for i in 1 : length(x)]
+create_autodiff(x, dx) = [ForwardDiff.Dual(x[i], dx[i]) for i in 1 : length(x)] # TODO: just use broadcast
 
 include("test_graph.jl")
 include("test_custom_collections.jl")