* add `IndexDict` and `CacheIndexDict`, new `Vector`-backed associative types that don't store the keys (indices) as they are directly constructible from the `Int` used to index into the values.
* make `BodyDict` and `JointDict` aliases for `IndexDict{BodyID}` and `IndexDict{JointID}`, stop using `UnsafeFastDict`. You can still index into e.g. `BodyDict` using a `RigidBody` as a convenience, but the key type is really `BodyID`
* change `Associative{<:RigidBody, T}` arguments in mechanism algorithms to `Associative{BodyID, T}` so that a `BodyDict` can be used.
* store the predecessor and successor ID for each joint ID in MechanismState upon construction; allows faster lookup of body data and parent body data in `BodyDict`s than going through the `Mechanism`'s `SpanningTree`.

A bit faster:
* `momentum_matrix`: 7.339 μs -> 6.545 μs
* `inverse_dynamics` 11.307 μs -> 10.657 μs
* `gravitational_potential_energy`: 2.787 μs -> 2.638 μs
* `dynamics`: 44.906 μs -> 40.151 μs
* `geometric_jacobian`: 4.213 μs -> 3.908 μs
* `mass_matrix`: 12.533 μs -> 9.671 μs